[feature] 분석 결과 Pie Chart api 연결

+++ core/data/src/main/java/com/ds/studify/core/data/datasource/StudyDataSource.kt
@@ -13,6 +13,6 @@ class StudyDataSource @Inject constructor(
     suspend fun postRecord(request: RequestRecordDto): BaseResponse<JsonElement?> =
         studyService.postRecord(request = request)
 
-    suspend fun getAnalysis(studyRecordId: Int): BaseResponse<ResponseAnalysisDto> =
+    suspend fun getAnalysis(studyRecordId: Long): BaseResponse<ResponseAnalysisDto> =
         studyService.getAnalysis(studyRecordId)
 }
